Move inset square creation
